refactor: [#4] 컨트롤러 통합.
MainController -> UserController
UserController에서 로그인과 회원 가입 매핑 처리.
