+++ QuantumTec/src/main/java/com/project/quantumtec/Controller/UserController.java
@@ -1,33 +1,29 @@
-package com.project.quantumtec.Controller.sub;
+package com.project.quantumtec.Controller;
 
 import com.project.quantumtec.DTO.user.LoginRequestDTO;
 import com.project.quantumtec.DTO.user.LoginResponseDTO;
-import com.project.quantumtec.DTO.user.UserDTO;
 import com.project.quantumtec.DTO.user.singupEmailCodeDTO;
 import com.project.quantumtec.Service.user.UserService;
 import com.project.quantumtec.VO.user.UserVO;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
+@RequestMapping("/user")
 public class UserController {
 
     @Autowired
     private UserService userService;
 
     // UserService 의 login 메소드를 호출하여 회원정보를 반환하는 메소드
+    @PostMapping("/login")
     public LoginResponseDTO login(LoginRequestDTO user) throws Exception {
         LoginResponseDTO loginResponseDTO = userService.login(user.getUserID(), user.getUserPW());
         return loginResponseDTO;
-//        if(loginResponseDTO != null) {
-//            return loginResponseDTO;
-//        }else {
-//            return null;
-//        }
     }
 
+    // UserService 의 입력받은 회원정보를 DB에 저장하는 메소드
+    @PostMapping("/signup")
     public UserVO signupAdd(@RequestBody UserVO user) throws Exception {
         UserVO checkUser = userService.signup(user);
         if(checkUser != null) {
@@ -37,20 +33,28 @@ public class UserController {
         }
     }
 
+    // UserService 의 입력받은 회원정보(아이디)가 중복되는지 확인하는 메소드
+    @PostMapping("/signup/checkid")
     public boolean checkDuplicateId(@RequestBody UserVO user) throws Exception {
         // true: 중복, false: 중복아님
         return userService.checkDuplicateId(user);
     }
 
+    // UserService 의 입력받은 회원정보(닉네임)가 중복되는지 확인하는 메소드
+    @PostMapping("/signup/checknickname")
     public boolean checkDuplicateNickname(@RequestBody UserVO user) throws Exception {
         // true: 중복, false: 중복아님
         return userService.checkDuplicateNickname(user);
     }
 
+    // UserService 의 입력받은 회원정보(이메일)에 인증코드를 전송하는 메소드
+    @PostMapping("/signup/send-email-auth")
     public void sendEmailAuth(@RequestBody UserVO user) throws Exception {
         userService.sendEmailAuth(user);
     }
 
+    // UserService 의 입력받은 회원정보(이메일)에 인증코드가 맞는지 확인하는 메소드
+    @PostMapping("/signup/check-email-auth")
     public boolean checkEmailAuth(@RequestBody singupEmailCodeDTO key) throws Exception {
         return userService.checkEmailAuth(key);
     }
